feat: change script for task 2 and add new prototype in lists.h dans all include inside .h

+++ singly_linked_lists/lists.h
@@ -1,7 +1,9 @@
 #ifndef LISTS_H
 #define LISTS_H
 
+#include <stdlib.h>
 #include <stddef.h>
+#include <string.h>
 
 /**
  * struct list_s - singly linked list
@@ -11,10 +13,11 @@
  *
  * Description: singly linked list node structure
  */
+
 typedef struct list_s
 {
 	char *str;
-	unsigned int len;
+	size_t len;
 	struct list_s *next;
 } list_t;
 
@@ -23,6 +26,5 @@ size_t list_len(const list_t *h);
 list_t *add_node(list_t **head, const char *str);
 list_t *add_node_end(list_t **head, const char *str);
 void free_list(list_t *head);
-
 
 #endif
